+++ NEWS.md
@@ -1,6 +1,9 @@
 # luz (development version)
 
 * Added support for arm Mac's and the MPS device. (#104)
+* Refactor checkpointing in luz - we now also serialize optimizer state and callbacks state. (#107)
+* Added a `luz_callback_autoresume()` allowing to easily resume trainining runs that might have crashed. (#107)
+* Added th `luz_callback_resume_from_checkpoint()` allowing one to resume a training run from a checkpoint file. (#107)
 
 # luz 0.3.1